impl(sidekick): move sidekick functionality to a package
The following series of commits intends to setup the following structure
for sidekick, per https://go.dev/doc/modules/layout:

cmd/
    sidekick/
        main.go
        main_test.go

internal/
    sidekick/
        ...files

This commit moves the sidekick logic to its own package.
