+++ generator/internal/sidekick/refreshall_test.go
@@ -1,0 +1,28 @@
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package sidekick
+
+import "testing"
+
+func TestRefreshAll(t *testing.T) {
+	cmdLine := &CommandLine{
+		Command:     "refreshall",
+		ProjectRoot: ".",
+		DryRun:      true,
+	}
+	if err := Root(cmdLine); err != nil {
+		t.Fatal(err)
+	}
+}
